resolve conflicts and separate test cases to better support after merge
